+++ src/main/java/com/web/UserEditor.java
@@ -4,6 +4,7 @@ import com.domain.User;
 import org.springframework.stereotype.Component;
 
 import java.beans.PropertyEditorSupport;
+import java.math.BigInteger;
 
 @Component
 public class UserEditor extends PropertyEditorSupport {
@@ -12,7 +13,7 @@ public class UserEditor extends PropertyEditorSupport {
     public void setAsText(String text) {
         try {
             User user = new User();
-            user.setUserId(Integer.parseInt(text));
+            user.setUserId(BigInteger.valueOf(Integer.parseInt(text)));
             this.setValue(user);
         } catch (Exception e) {
             e.printStackTrace();
